Don't use a slice for the word field in the Word structure.
Rust started panicking for unaligned pointers in
core::slice::from_raw_parts.  The Word structure could move to using a
Vec for the word field and have a set of serialization /
deserialization methods.  That will require a large rewrite.

Instead this commit changes to using a raw const pointer in the meantime.

There are some other issues with clippy warnings and unnecessary mut
pointers that need to also be fixed up in another commit.